+++ include/async_mqtt5/impl/subscribe_op.hpp
@@ -92,10 +92,8 @@ public:
 			topics, props
 		);
 
-		auto max_packet_size = static_cast<size_t>(
-			_svc_ptr->connack_property(prop::maximum_packet_size)
-				.value_or(default_max_send_size)
-		);
+		auto max_packet_size = _svc_ptr->connack_property(prop::maximum_packet_size)
+				.value_or(default_max_send_size);
 		if (subscribe.size() > max_packet_size)
 			return complete_post(client::error::packet_too_large, packet_id);
